Split namespaces into separate files
