+++ types/trigger.d.ts
@@ -1,0 +1,79 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the MIT License.
+
+import { CosmosDBTrigger, CosmosDBTriggerOptions } from './cosmosDB';
+import { EventGridTrigger, EventGridTriggerOptions } from './eventGrid';
+import { EventHubTrigger, EventHubTriggerOptions } from './eventHub';
+import { GenericTriggerOptions } from './generic';
+import { HttpTrigger, HttpTriggerOptions } from './http';
+import { FunctionTrigger } from './index';
+import {
+    ServiceBusQueueTrigger,
+    ServiceBusQueueTriggerOptions,
+    ServiceBusTopicTrigger,
+    ServiceBusTopicTriggerOptions,
+} from './serviceBus';
+import {
+    StorageBlobTrigger,
+    StorageBlobTriggerOptions,
+    StorageQueueTrigger,
+    StorageQueueTriggerOptions,
+} from './storage';
+import { TimerTrigger, TimerTriggerOptions } from './timer';
+
+/**
+ * The root namespace used to help create trigger configuration (the primary input)
+ * You can create trigger config without this namespace, but it provides features like autocomplete, better build errors, and it will set the `name` property for you
+ */
+export namespace trigger {
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-http-webhook-trigger?&pivots=programming-language-javascript)
+     */
+    export function http(options: HttpTriggerOptions): HttpTrigger;
+
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-timer?pivots=programming-language-javascript)
+     */
+    export function timer(options: TimerTriggerOptions): TimerTrigger;
+
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-storage-blob-trigger?pivots=programming-language-javascript)
+     */
+    export function storageBlob(options: StorageBlobTriggerOptions): StorageBlobTrigger;
+
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-storage-queue-trigger?pivots=programming-language-javascript)
+     */
+    export function storageQueue(options: StorageQueueTriggerOptions): StorageQueueTrigger;
+
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-service-bus-trigger?pivots=programming-language-javascript)
+     */
+    export function serviceBusQueue(options: ServiceBusQueueTriggerOptions): ServiceBusQueueTrigger;
+
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-service-bus-trigger?pivots=programming-language-javascript)
+     */
+    export function serviceBusTopic(options: ServiceBusTopicTriggerOptions): ServiceBusTopicTrigger;
+
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-event-hubs-trigger?pivots=programming-language-javascript)
+     */
+    export function eventHub(options: EventHubTriggerOptions): EventHubTrigger;
+
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-event-grid-trigger?pivots=programming-language-javascript)
+     */
+    export function eventGrid(options: EventGridTriggerOptions): EventGridTrigger;
+
+    /**
+     * [Link to docs and examples](https://docs.microsoft.com/azure/azure-functions/functions-bindings-cosmosdb-v2-trigger?pivots=programming-language-javascript)
+     */
+    export function cosmosDB(options: CosmosDBTriggerOptions): CosmosDBTrigger;
+
+    /**
+     * A generic option that can be used for any trigger type
+     * Use this method if your desired trigger type does not already have its own method
+     */
+    export function generic(options: GenericTriggerOptions): FunctionTrigger;
+}
